DAS-2067 - Migrate datatree_render.py.
