+++ xarray/core/datatree.py
@@ -23,6 +23,7 @@ from xarray.core.datatree_mapping import (
     check_isomorphic,
     map_over_subtree,
 )
+from xarray.core.datatree_render import RenderDataTree
 from xarray.core.formatting import datatree_repr
 from xarray.core.formatting_html import (
     datatree_repr as datatree_repr_html,
@@ -46,7 +47,6 @@ from xarray.datatree_.datatree.ops import (
     MappedDatasetMethodsMixin,
     MappedDataWithCoords,
 )
-from xarray.datatree_.datatree.render import RenderTree
 
 try:
     from xarray.core.variable import calculate_dimensions
@@ -1451,7 +1451,7 @@ class DataTree(
 
     def render(self):
         """Print tree structure, including any data stored at each node."""
-        for pre, fill, node in RenderTree(self):
+        for pre, fill, node in RenderDataTree(self):
             print(f"{pre}DataTree('{self.name}')")
             for ds_line in repr(node.ds)[1:]:
                 print(f"{fill}{ds_line}")
